Some cleaning up and new features
You can now trigger bells manually (just click the Trigger link in the
web UI), plus the update now has a warning screen. Monitor page also
included, which is a handy page for small screens attached to the Pi
(such as the PiTFT). Shows total loaded bells and the time of the next
bell.